+++ e2e/ts/jest.config.js
@@ -1,5 +1,6 @@
-module.exports = {
+export default {
+  extensionsToTreatAsEsm: ['.ts'],
   transform: {
-    '^.+\\.ts$': 'ts-jest',
+    '^.+\\.ts$': ['ts-jest', { useESM: true }],
   },
 };